+++ packages/riverpod/CHANGELOG.md
@@ -11,12 +11,13 @@
 - The `Reader` typedef is removed. Use `Ref` instead.
 - `ProviderListener` is removed. Used `ref.listen` instead.
 
-Non breaking changes:
+Non-breaking changes:
 
 - feat: A new `AutoDisposeRef.keepAlive()` function is added. It is meant to replace
   `AutoDisposeRef.maintainState` to make logic for preventing the disposal of a provider more reusable.
 - `AutoDisposeRef.maintainState` is deprecated. Use the new ``AutoDisposeRef.keepAlive()` instead.
 - feat: Add support for `ref.invalidate(family)` to recompiute an entire family (#1517)
+- fixed a bug where `AsyncValue.whenData` did not preserve `AsyncValue.isLoading/isRefreshing`
 
 # 2.0.0-dev.9